refactor(payments): move Polar Better Auth plugins out of auth.server
Keep auth config focused on core auth and own Polar plugin wiring in payments.
